requirements.txt
# =============================================================================
# SMART LAMP PROJECT REQUIREMENTS
# VIP Project Group E - Spring 2025
# =============================================================================

# Core Python packages
asyncio-mqtt==0.16.1
python-dotenv==1.0.0

# Raspberry Pi GPIO and Hardware
RPi.GPIO==0.7.1
spidev==3.6
adafruit-circuitpython-mcp3xxx==1.4.13
rpi-ws281x==4.3.4

# Web Framework and UI
streamlit==1.28.1
plotly==5.17.0
fastapi==0.104.1
uvicorn==0.24.0

# HTTP Requests and APIs
requests==2.31.0
aiohttp==3.9.1
urllib3==2.1.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2

# Audio Processing
pygame==2.5.2
python-vlc==3.0.18121

# Database
sqlalchemy==2.0.23

# Data Processing and Analysis
scipy==1.11.4
matplotlib==3.8.2

# Utilities and Helpers
colorama==0.4.6
pydantic==2.5.2
typing-extensions==4.8.0
click==8.1.7

# Date and Time Handling
python-dateutil==2.8.2

# JSON and Data Validation
jsonschema==4.20.0

# Logging and Monitoring
structlog==23.2.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0